refactor(mjh/job_analysis): split job_analysis_service.py (786 LOC) into analyze + promote modules
- job_analysis_service.py: analyze/score/get_analysis/soft_delete_analysis + all validation/prompt helpers (~470 LOC)
- job_analysis_promote_service.py: apply_to_application + _find_or_create_company (~115 LOC)
- _job_analysis_utils.py: shared tiny utilities (_str_or_none, _safe_float, _safe_remote_type, _map_salary_period) to break the circular import between the two sibling modules (~50 LOC)

apply_to_application re-exported from job_analysis_service for backward compat — no caller changes required.
37/37 tests pass.
